update national_id type to bigent

+++ Eskan_Enterprise/eskan-enterprise/database/migrations/2022_05_31_141646_add_national_id_to_customers_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
     public function up()
     {
         Schema::table('customers', function (Blueprint $table) {
-            $table->integer('national_id');
+            $table->bigInteger('national_id');
         });
     }
